- Add optIn to UserInfo.kt.  - Moved UpdateViewModel.kt to settings path.  - Remove Experimental annotations.
