+++ app/src/main/java/com/shov/unlimstorage/viewModels/navigations/MainNavigationViewModel.kt
@@ -10,18 +10,18 @@ import com.shov.unlimstorage.values.UNCHECKED_CAST
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 
-class MainNavigationViewModel @ExperimentalMaterialApi @AssistedInject constructor(
+class MainNavigationViewModel @OptIn(ExperimentalMaterialApi::class) @AssistedInject constructor(
 	@Assisted private val _scaffoldState: ScaffoldState,
 	@Assisted private val _sheetState: ModalBottomSheetState
 ) : ViewModel() {
 	val scaffoldState get() = _scaffoldState
 
-	@ExperimentalMaterialApi
+	@OptIn(ExperimentalMaterialApi::class)
 	val sheetState = _sheetState
 
 	@Suppress(UNCHECKED_CAST)
 	companion object {
-		@ExperimentalMaterialApi
+		@OptIn(ExperimentalMaterialApi::class)
 		fun provideFactory(
 			assistedFactory: MainNavigationViewModelFactory,
 			scaffoldState: ScaffoldState,
