+++ app/src/main/java/com/shov/unlimstorage/views/navigations/MainNavigation.kt
@@ -1,7 +1,6 @@
 package com.shov.unlimstorage.views.navigations
 
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
@@ -11,18 +10,13 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import coil.annotation.ExperimentalCoilApi
 import com.shov.unlimstorage.models.preferences.Preference
 import com.shov.unlimstorage.values.IS_AUTH
 import com.shov.unlimstorage.values.Screen
 import com.shov.unlimstorage.viewModels.MainNavigationViewModel
 import com.shov.unlimstorage.views.SignInScreen
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 
-@ExperimentalCoilApi
-@ExperimentalFoundationApi
-@ExperimentalCoroutinesApi
-@ExperimentalMaterialApi
+@OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun MainNavigation(
 	mainNavigationViewModel: MainNavigationViewModel,
