+++ app/src/main/java/com/shov/unlimstorage/views/settings/updates/UpdateScreen.kt
@@ -14,7 +14,7 @@ import androidx.navigation.NavController
 import com.alorma.compose.settings.ui.SettingsMenuLink
 import com.shov.unlimstorage.R
 import com.shov.unlimstorage.viewModels.TopAppBarViewModel
-import com.shov.unlimstorage.viewModels.UpdateViewModel
+import com.shov.unlimstorage.viewModels.settings.UpdateViewModel
 
 @Composable
 fun UpdateScreen(
